+++ services/orchestration/decision-engine.js
@@ -1,0 +1,9 @@
+const { buildResponsePlan } = require('./response-planner');
+
+function decideOrchestration(input = {}) {
+  return buildResponsePlan(input);
+}
+
+module.exports = {
+  decideOrchestration,
+};