+++ services/orchestration/index.js
@@ -1,0 +1,9 @@
+const { decideOrchestration } = require('./decision-engine');
+const { detectPriorityIntent } = require('./intent-hierarchy');
+const { buildResponsePlan } = require('./response-planner');
+
+module.exports = {
+  buildResponsePlan,
+  decideOrchestration,
+  detectPriorityIntent,
+};